Add more context and example for Box JWT
The added examples work as of Box 2.0, use Cygwin to run the highlighted commands. (I generated my own private/public key pair because the one generated in Box is difficult to use on salesforce as it is generated as encrypted while salesforce only work with decrypted private keys).
To use the userId type of authentication, make sure Application Access in Developer console is set to Entreprise.